new IOC version aware CI

+++ helm-push.sh
@@ -16,43 +16,48 @@ fi
 
 # Helm Registry in GHCR
 HELM_REPO=ghcr.io/epics-containers
-HELM_OCI=oci://${HELM_REPO}
 
 # extract name from the Chart.yaml
-NAME=$(awk '/^name:/{print $NF}' "${IOC_ROOT}/Chart.yaml")
-CHART=${HELM}/${NAME}
-CHART_OCI=${HELM_OCI}/${NAME}
+NAME=$(sed -n '/^name: */s///p' "${IOC_ROOT}/Chart.yaml")
+CHART=oci://${HELM_REPO}/${NAME}
 
 # Before calling this script: set CR_PAT to a github personal access token 
 # see https://github.com/settings/tokens
 echo $CR_PAT | helm registry login -u USERNAME --password-stdin $HELM_REPO
 
-echo "deploying ${IOC_ROOT} to helm repo as version ${TAG}"
+echo "Verifying deploy of ${IOC_ROOT} to helm repo as version ${TAG} ..."
 helm package "${IOC_ROOT}" -u --app-version ${TAG} --version ${TAG}
 PACKAGE=$(realpath ${NAME}-${TAG}.tgz)
 
-# determine if this IOC has changed since latest helm chart push
-TMPDIR=$(mktemp -d)
-cd $TMPDIR
-helm pull ${CHART_OCI} > out.txt
-LATEST_VERSION=$(sed -n '/^Pulled/s=.*:==p' out.txt)
-echo latest chart: $LATEST_VERSION
-mkdir latest_ioc this_ioc
-tar -xf *tgz -C latest_ioc
-# repackage the new IOC with same version as above for direct comparison
-# (packaging reformats the Chart.yaml file so this is the best approach)
-helm package "${IOC_ROOT}" -u --app-version ${LATEST_VERSION} --version ${LATEST_VERSION}
-tar -xf *tgz -C this_ioc
-echo comparing $(realpath latest) $(realpath this) 
-if diff -r latest_ioc this_ioc &> diff.out ; then 
-    MATCH=YES ; else MATCH=NO
-fi
-cd -
-rm -r $TMPDIR
+# Temp dir for testing if the helm chart has changed
+TMPDIR=$(mktemp -d); cd ${TMPDIR}
 
-if [ "$MATCH" == "NO" ] ; then    
-    helm push "${PACKAGE}" ${HELM_OCI}
-else 
-    echo "not pushing unchanged IOC"
+# extract the latest version to a temporary folder
+helm pull ${CHART} > out.txt
+LATEST_VERSION=$(sed -n '/^Pulled:/s/.*://p' out.txt)
+echo "latest version of ${CHART} is ${LATEST_VERSION}"
+mkdir latest_ioc this_ioc
+tar -xf *tgz -C latest_ioc &> tar1.out
+
+# repackage the new IOC with same version as above for direct comparison
+# (packaging reformats the Chart.yaml file so this is the simplest approach)
+helm package "${IOC_ROOT}" --app-version ${LATEST_VERSION} --version ${LATEST_VERSION}
+tar -xf *tgz -C this_ioc &> tar2.out
+
+# compare the packages and push the new package if it has changed
+if diff -r latest_ioc this_ioc &> diff.out ; then 
+    echo "not pushing unchanged IOC ${CHART} version ${LATEST_VERSION}"   
+else
+    echo "pushing ${CHART} version ${TAG}"   
+    helm push "${PACKAGE}" oci://${HELM_REPO}
 fi
-rm "${PACKAGE}"
+
+cat diff.out
+
+# clean up
+if [ -z ${HELM_PUSH_DEBUG} ] ; then
+    rm "${PACKAGE}"
+    rm -r ${TMPDIR}
+else
+    echo "helm-push.sh comparison output is in ${TMPDIR}"
+fi
